Moved OpenAPI endpoint examples classes and helper methods to new openapi package. Updated example data.
